fixup! 通知ボタン押せる押せないを変更
