+++ src/app/account/page.tsx
@@ -1,22 +1,16 @@
 "use client";
 import { auth } from "@/app/firebase";
-import {
-  requestPermission,
-  revokePermission,
-} from "@/utils/CloudMessaging/notificationController";
+import Notification from "@/Components/Notification/Notification";
 import { createUser } from "@/utils/createUserAuth";
 import { loginUser } from "@/utils/loginUserAuth";
 import { signInAsGuest } from "@/utils/signInAnonymously";
 import { signInWithGoogleAccount } from "@/utils/signInWithGoogleAccount";
 import { useUser } from "@/utils/UserContext";
-import Alert from "@mui/material/Alert";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import MuiCard from "@mui/material/Card";
-import CircularProgress from "@mui/material/CircularProgress";
 import CssBaseline from "@mui/material/CssBaseline";
 import Divider from "@mui/material/Divider";
-import Snackbar from "@mui/material/Snackbar";
 import Stack from "@mui/material/Stack";
 import { styled } from "@mui/material/styles";
 import TextField from "@mui/material/TextField";
@@ -24,7 +18,7 @@ import ToggleButton from "@mui/material/ToggleButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import Typography from "@mui/material/Typography";
 import { signOut } from "firebase/auth";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 const Card = styled(MuiCard)(({ theme }) => ({
   display: "flex",
@@ -58,12 +52,6 @@ export default function UserForm() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [formMode, setFormMode] = useState<"register" | "login">("register");
-  const [notificationTokenGenerating, setNotificationTokenGenerating] =
-    useState(false);
-  const [notificationMessage, setNotificationMessage] = useState<string | null>(
-    null
-  );
-  const [isNotificationActive, setIsNotificationActive] = useState(false);
 
   const { user } = useUser();
 
@@ -102,23 +90,6 @@ export default function UserForm() {
     if (newMode) setFormMode(newMode);
   };
 
-  const handleShowNotification = (message: string) => {
-    setNotificationMessage(message);
-  };
-
-  // ページ読み込み時に Service Worker 状態を確認
-  useEffect(() => {
-    if ("serviceWorker" in navigator) {
-      navigator.serviceWorker.ready
-        .then((registration) => {
-          setIsNotificationActive(!!registration.active); // active が存在するかで判定
-        })
-        .catch((error) => {
-          console.error("Error checking ServiceWorker status:", error);
-        });
-    }
-  }, []);
-
   return (
     <>
       <CssBaseline enableColorScheme />
@@ -151,49 +122,7 @@ export default function UserForm() {
             {user ? (
               <>
                 <Typography>ログイン中: {user.name}</Typography>
-                <Box sx={{ display: "flex", justifyContent: "center", gap: 2 }}>
-                  <Button
-                    variant="contained"
-                    onClick={() => {
-                      setNotificationTokenGenerating(true);
-                      requestPermission(() => {
-                        setNotificationTokenGenerating(false);
-                        handleShowNotification("通知を受信しました");
-                        setIsNotificationActive(true); // 通知状態を有効化
-                      });
-                    }}
-                    disabled={
-                      notificationTokenGenerating || isNotificationActive
-                    }
-                    startIcon={
-                      notificationTokenGenerating ? (
-                        <CircularProgress size={20} />
-                      ) : null
-                    }
-                  >
-                    通知を受信
-                  </Button>
-
-                  <Button
-                    variant="contained"
-                    onClick={() => {
-                      revokePermission();
-                      handleShowNotification("通知を解除しました");
-                      setIsNotificationActive(false); // 通知状態を無効化
-                    }}
-                    disabled={!isNotificationActive}
-                  >
-                    通知を解除
-                  </Button>
-                  <Snackbar
-                    open={!!notificationMessage}
-                    anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
-                    autoHideDuration={3000} // 3秒で自動的に非表示
-                    onClose={() => setNotificationMessage(null)} // 非表示時にメッセージをリセット
-                  >
-                    <Alert severity="info">{notificationMessage}</Alert>
-                  </Snackbar>
-                </Box>
+                {user.loginType !== "Guest" && <Notification />}
                 <RoundedButton variant="contained" onClick={handleLogout}>
                   ログアウト
                 </RoundedButton>
